Move coordinator.Result to spi
