+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/maintenance/TemporaryBuildsCleanerAsyncInvoker.java
@@ -20,6 +20,7 @@ package org.jboss.pnc.coordinator.maintenance;
 import org.jboss.pnc.common.concurrent.NamedThreadFactory;
 import org.jboss.pnc.model.BuildConfigSetRecord;
 import org.jboss.pnc.model.BuildRecord;
+import org.jboss.pnc.spi.coordinator.Result;
 import org.jboss.pnc.spi.datastore.repositories.BuildConfigSetRecordRepository;
 import org.jboss.pnc.spi.datastore.repositories.BuildRecordRepository;
 import org.jboss.pnc.spi.exception.ValidationException;
