+++ rest/src/main/java/org/jboss/pnc/rest/endpoint/BuildRecordEndpoint.java
@@ -17,12 +17,14 @@
  */
 package org.jboss.pnc.rest.endpoint;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.jboss.pnc.coordinator.maintenance.Result;
+import org.jboss.pnc.common.util.HttpUtils;
+import org.jboss.pnc.spi.coordinator.Result;
 import org.jboss.pnc.coordinator.maintenance.TemporaryBuildsCleanerAsyncInvoker;
 import org.jboss.pnc.model.BuildRecord;
 import org.jboss.pnc.model.BuildStatus;
