+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/Result.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.coordinator.maintenance;
+package org.jboss.pnc.spi.coordinator;
 
 import lombok.Getter;
